module 3. Final

+++ src/main/java/lang/print/gaps/finalModuleTask/TemperatureConverter.java
@@ -1,7 +1,7 @@
 package lang.print.gaps.finalModuleTask;
 
 public class TemperatureConverter {
-    public void toFahrenheit(int temperatureCelsius) {
+    public void toFahrenheit(double temperatureCelsius) {
         double f = temperatureCelsius * 9 / 5 + 32;
         System.out.println(f);
     }
